unconvert: avoid panicking on out-of-bounds column values
For example:

    println(C.int(0), C.int(0), C.int(0), int(int(0)))

is rewritten by cgo into:

    println(_Ctype_int(0), _Ctype_int(0), _Ctype_int(0), int(int(0)))

The column location for the unnecessary int conversion in the
rewritten form is out-of-bounds for the prior form and would cause a
panic.

Ideally we would either not print a caret at all in this case or
somehow map the rewritten file's column number back to the source
file's column number, but there doesn't seem to be any
simple/satisfying way to do that at the moment.