unconvert: add experimental -configs flag
With this, users can supply a list of build configs (each specified as
a list of extra key=val environment variables to set) under which to
run unconvert. Unconvert will then load the packages under all
configs, and use the same logic as -all to merge them.

For example, on a linux/amd64 host with gcc cross-compilers for
linux/386, you might use

    unconvert -configs='[[], ["GOARCH=386", "CGO_ENABLED=1", "CC_FOR_TARGET=gcc -m32"]]' mypkgs...

to test both the native build and linux/386 build with cgo support.

Updates #26.