unconvert: fix -apply for cgo sources
go/packages gives us a path to the cgo-processed source files (which
is in the build cache and probably bad to modify), but we want to
rewrite the original source files. As a quick hack, take the file name
of the package token's position.

Also, refactor editSet to clear the Offset, since we only care about
file/name/column, and it seems to be causing problems.

Fixes #38.